Декомпозирует файл типов по категориям
