+++ src/pages/favorites-page/components/favorites-item/favorites-item.tsx
@@ -2,11 +2,9 @@ import { useAppDispatch } from '../../../../hooks';
 import CardsList from '../../../../components/cards-list/cards-list';
 import LocationsItemLink from '../../../../components/locations-item-link/locations-item-link';
 import { changeCity } from '../../../../store/actions';
-import {
-  CityKeys,
-  ShortOfferListType,
-  TypesPageKeys,
-} from '../../../../types/types';
+import { TypesPageKeys } from '../../../../types/types';
+import { ShortOfferListType } from '../../../../types/offers';
+import { CityKeys } from '../../../../types/cities';
 
 type FavoritesItemProps = {
   typesPage: TypesPageKeys;
