+++ src/pages/favorites-page/favorites-page.tsx
@@ -5,10 +5,11 @@ import Nav from '../../components/nav/nav';
 import Logo from '../../components/logo/logo';
 import { TypesPage } from '../../const';
 import { Title } from '../../components/title/title';
-import { FavoritesListType, TypesPageKeys } from '../../types/types';
+import { TypesPageKeys } from '../../types/types';
 import { groupByList } from './utils';
 import { FavoriteEmpty } from './components/favorites-empty/favorites-empty';
 import { FavoritesList } from './components/favorites-list/favorites-list';
+import { FavoritesListType } from '../../types/favorites';
 
 type FavoritePageProps = {
   favorites: FavoritesListType;
