+++ src/store/actions.ts
@@ -1,5 +1,7 @@
 import { createAction } from '@reduxjs/toolkit';
-import { CityKeys, ShortOfferListType, SortTypeKeys } from '../types/types';
+import { SortTypeKeys } from '../types/types';
+import { CityKeys } from '../types/cities';
+import { ShortOfferListType } from '../types/offers';
 
 const changeCity = createAction<CityKeys>('location/changeCity');
 
@@ -7,8 +9,4 @@ const setOffers = createAction<ShortOfferListType>('offers/setOffers');
 
 const changeSortKey = createAction<SortTypeKeys>('sort/changeSortKey');
 
-export {
-  changeCity,
-  setOffers,
-  changeSortKey,
-};
+export { changeCity, setOffers, changeSortKey };
