+++ src/types/favorites.ts
@@ -1,0 +1,9 @@
+import { OfferType, ShortOfferListType } from './offers';
+
+type FavoritesType = OfferType & { previewImage: string };
+
+type FavoritesListType = FavoritesType[];
+
+type GroupedOffersType = Record<string, ShortOfferListType>;
+
+export type { GroupedOffersType, FavoritesListType };
