+++ src/types/map.ts
@@ -1,0 +1,3 @@
+import { OfferType } from './offers';
+
+export type MapPointsListType = Pick<OfferType, 'id' | 'city' | 'location'>[];
